# Add copyto to facilitate broadcasting