This PR adds a custom broadcast style for `AbstractKroneckerProducts` which uses the efficient `collect!` wherever possible to evaluate broadcast operations without indexing into the Kronecker product (which is slow). This often leads to significant speed-ups.

On master
```julia
julia> K1 = kronecker(ones(30, 30), ones(30, 30));

julia> K2 = kronecker(ones(30, 30), ones(30, 30));

julia> @btime $K1 + $K2;
  56.020 ms (2 allocations: 6.18 MiB)

julia> @btime $K1 .+ $K2;
  55.951 ms (2 allocations: 6.18 MiB)

julia> @btime -$K1;
  28.102 ms (2 allocations: 6.18 MiB)

julia> @btime copy($K1);
  27.994 ms (2 allocations: 6.18 MiB)
```

The timing is dominated entirely by indexing

```julia
julia> @btime $K1[1,1];
  34.343 ns (0 allocations: 0 bytes)

julia> 34.343 * 1e-6 * length(K1) # time in ms to loop over the array elementwise
27.81783
```

After this PR

```julia
julia> @btime $K1 + $K2;
  1.503 ms (2 allocations: 6.18 MiB)

julia> @btime $K1 .+ $K2;
  1.491 ms (2 allocations: 6.18 MiB)

julia> @btime -$K1;
  1.047 ms (2 allocations: 6.18 MiB)

julia> @btime copy($K1);
  735.405 μs (2 allocations: 6.18 MiB)
```

The speed of the sum is comparable to that of dense arrays now:

```julia
julia> @btime $(collect(K1)) + $(collect(K2));
  1.714 ms (2 allocations: 6.18 MiB)
```