@MichielStock 2021-09-19
Requested and scheduled!

> On 19 Sep 2021, at 18:35, Jishnu Bhattacharya ***@***.***> wrote:
> 
> 
> Could you tag a new release? Thanks!
> 
> —
> You are receiving this because you modified the open/close state.
> Reply to this email directly, view it on GitHub <https://github.com/MichielStock/Kronecker.jl/pull/100#issuecomment-922501486>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/ABN5PI54NNBOFFIY3FVYN43UCYGL3ANCNFSM5EIAXUFA>.
>